chore: add typings file
