+++ client/plugin/views/MyOrdersView.tsx
@@ -1,22 +1,15 @@
 import { Transaction } from "ethers";
 import { FunctionalComponent, h } from "preact";
-import { StateUpdater, useState } from "preact/hooks";
+import { useState } from "preact/hooks";
 import { Button } from "../components/Button";
 import { Listing, Order } from "../components/MyListingsContext";
-import { ActiveSigner } from "../components/SignerContext";
 import { getListingsWithOrdersFromAddress } from "../helpers/transactions";
 import { useContract } from "../hooks/use-contract";
 import { useListings } from "../hooks/use-listings";
 import { useSigner } from "../hooks/use-signer";
+import { OrderDetailsProps, ActiveSigner } from "../typings/typings";
 import { listingStyles, orderStyles } from "./MyListingsView";
 
-
-type OrderDetailsProps = {
-	order: Order;
-	action: (() => void) | (() => Promise<Transaction>);
-	childrenAction: string;
-	buttonDisabled: boolean;
-};
 
 export const OrderDetails: FunctionalComponent<OrderDetailsProps> = (props: OrderDetailsProps) => {
 	return (
@@ -40,7 +33,7 @@ export function MyOrdersView () {
 	const refund = async (listing: Listing, order: Order): Promise<Transaction> => {
 		const tx = await market.refund(listing.listingId, order.orderId);
 		console.log("refund tx: ", tx);
-		return tx
+		return tx;
 	};
 
 	return (
@@ -50,7 +43,7 @@ export function MyOrdersView () {
 				{myOrders.map((listing) => (
 					<div style={{ display: "grid", rowGap: "4px" }}>
 						{listing.orders.map((order) => (
-							<OrderDetails order={order} action={() => refund(listing.listing, order)} childrenAction={'refund'} buttonDisabled={!order.isActive} />
+							<OrderDetails order={order} action={async () => await refund(listing.listing, order)} childrenAction={'refund'} buttonDisabled={!order.isActive} />
 						)
 						)}
 					</div>
